Prepare HtmlTinkerX 3.0.1 release
