+++ Sources/PSParseHTML.PowerShell/PSParseHTML.PowerShell.csproj
@@ -3,7 +3,7 @@
     <PropertyGroup>
         <Company>Evotec</Company>
         <Authors>Przemyslaw Klys</Authors>
-        <VersionPrefix>3.0.0</VersionPrefix>
+        <VersionPrefix>3.0.1</VersionPrefix>
         <TargetFrameworks>net472;net8.0</TargetFrameworks>
         <AssemblyName>PSParseHTML.PowerShell</AssemblyName>
         <AssemblyTitle>PSParseHTML.PowerShell</AssemblyTitle>
